fix: snippets moved to src/snippets
