--- .github/workflows/build.yml
@@ -1,59 +1,0 @@
-name: Yarn build
-
-on:
-  push:
-    branches: [ "master" ]
-  pull_request:
-    branches: [ "master" ]
-  schedule:
-    - cron: '20 2 * * 3'
-
-jobs:
-  build:
-    name: Compile and Lint Codebase
-    runs-on: ubuntu-latest
-    # permissions:
-    #   contents: read
-    #   security-events: write
-    #   actions: read
-    steps:
-      - name: Install Act dependencies
-        if: ${{ env.ACT }}
-        run: |
-          apt-get update && apt-get install sudo -y
-
-      # - name: Install Fish Shell
-      - uses: fish-actions/install-fish@v1.1.0
-      - uses: actions/checkout@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version: '20'
-          cache: 'yarn'
-      - run: yarn install --ignore-scripts
-      - run: yarn lint
-
-
-      # - name: Set Node.js 18.x
-      #   uses: actions/setup-node@v3
-      #   with:
-      #     node-version: 18.x
-      #
-      # - name: Install Dependencies
-      #   uses: borales/actions-yarn@v5
-      #   with:
-      #     cmd: install
-      #   env:
-      #     NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
-      #
-      # - name: List Files in Workspace
-      #   run: ls -la
-      #
-      # - name: yarn Compile
-      #   uses: borales/actions-yarn@v5
-      #   with:
-      #     cmd: compile
-      #
-      # - name: Run ESLint
-      #   uses: borales/actions-yarn@v5
-      #   with:
-      #     cmd: lint
