--- .github/workflows/compile-and-link.yml
@@ -1,31 +1,0 @@
-name: Compile and Link TypeScript Command
-
-on:
-  pull_request:
-    branches:
-      - master
-
-jobs:
-  compile-and-link:
-    name: Compile and Link Command
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout Code
-        uses: actions/checkout@v3
-
-      - name: Install Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 21.7.1
-
-      - name: Install Yarn
-        run: npm install -g yarn@1.22.22
-
-      - name: Install Dependencies
-        run: yarn install
-
-      - name: Verify Command
-        run: |
-          echo "Checking if command is globally accessible..."
-          fish-lsp --help
