--- .github/workflows/debug.yml
@@ -1,54 +1,0 @@
-name: debug
-
-on:
-  pull_request:
-    branches:
-      - master
-
-jobs:
-  debug:
-    name: Lint Codebase
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout Code
-        uses: actions/checkout@v3
-
-      - name: Install Fish Shell
-        uses: fish-actions/install-fish@v1.1.0
-
-
-      - name: Install Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 21.7.1
-
-      - name: Run install
-        uses: Borales/actions-yarn@v5.0.0
-        with:
-          cmd: install --ignore-scripts # will run `yarn install` command
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }} # if needed
-
-      - name: Yarn Install sh setup
-        uses: Borales/actions-yarn@v5.0.0
-        with:
-          run: setup
-          dir: ./
-
-      - name: Verify Yarn Version
-        run: yarn --version
-
-      - name: Verify Node Version
-        run: node --version
-
-      #
-      # - name: List Node Modules
-      #   run: ls -la node_modules
-
-      # - name: Compile TypeScript
-      #   run: yarn tsc -b
-        # env:
-        #   DEBUG: "true"
-      # - name: Run ESLint
-      #   run: yarn lint
