--- .github/workflows/lint-testing.yml
@@ -1,41 +1,0 @@
-name: lint testing
-
-on:
-  push:
-    branches: [ "master" ]
-  pull_request:
-    branches: [ "master" ]
-  schedule:
-    - cron: '20 2 * * 3'
-
-jobs:
-  eslint:
-    name: Run ESLint Scanning
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      security-events: write
-      actions: read
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v3
-
-      - name: Install Fish Shell
-        uses: fish-actions/install-fish@v1.1.0
-
-      - name: Set up Node.js 18.x
-        uses: actions/setup-node@v3
-        with:
-          node-version: 18.x
-
-      - name: Run install
-        uses: borales/actions-yarn@v4
-        with:
-          cmd: install # will run `yarn install` command
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }} # if needed
-
-      - name: Build production bundle
-        uses: borales/actions-yarn@v4
-        with:
-          cmd: lint # will run `yarn build:prod` command
