+++ scripts/build-completions.fish
@@ -1,5 +1,3 @@
-#!/usr/bin/env fish
-
 if not test -d $HOME/.config/fish/completions
   mkdir -p $HOME/.config/fish/completions 
 end
