+++ src/snippets/fishlspEnvVariables.json
@@ -1,0 +1,52 @@
+[
+  {
+    "name": "fish_lsp_enabled_handlers",
+    "description": "enables the fish-lsp handlers 'asciiArt', 'formatting', 'logging', 'complete', 'hover', 'rename', 'definition', 'references', 'diagnostics', 'signatureHelp', 'codeAction', 'index'",
+    "valueType": "array"
+  },
+  {
+    "name": "fish_lsp_disabled_handlers",
+    "description": "disables the fish-lsp handlers 'asciiArt', 'formatting', 'logging', 'complete', 'hover', 'rename', 'definition', 'references', 'diagnostics', 'signatureHelp', 'codeAction', 'index'",
+    "valueType": "array"
+  },
+  {
+    "name": "fish_lsp_commit_characters",
+    "description": "array of the completion expansion characters. Single letter values only. Commit characters are used to select completion items, as shortcuts. (default: [])",
+    "valueType": "array"
+  },
+  {
+    "name": "fish_lsp_logfile",
+    "description": "path to the logs.txt file (default: '~/path/to/fish-lsp/logs.txt')",
+    "valueType": "string"
+  },
+  {
+    "name": "fish_lsp_format_tabsize",
+    "description": "amount of spaces in a tab character for the formatter provider (default: 4)",
+    "valueType": "number"
+  },
+  {
+    "name": "fish_lsp_format_switch_case",
+    "description": "keep case statements left aligned with switch block. (default: false)",
+    "valueType": "boolean"
+  },
+  {
+    "name": "fish_lsp_all_indexed_paths",
+    "description": "fish file paths to include as workspaces (default: ['/usr/share/fish', '$HOME/.config/fish'])",
+    "valueType": "array"
+  },
+  {
+    "name": "fish_lsp_modifiable_paths",
+    "description": "fish file paths that can be renamed by the user (default: ['$HOME/.config/fish'])",
+    "valueType": "array"
+  },
+  {
+    "name": "fish_lsp_diagnostic_disable_error_codes",
+    "description": "disable diagnostics for matching error codes (default: [])",
+    "valueType": "array"
+  },
+  {
+    "name": "fish_lsp_max_background_files",
+    "description": "maximum number of background files to read into buffer on startup",
+    "valueType": "number"
+  }
+]
