+++ src/snippets/specialFishVariables.json
@@ -1,0 +1,90 @@
+[
+  {
+    "name": "PATH",
+    "description": "A list of directories in which to search for commands. This is a common unix variable also used by other tools."
+  },
+  {
+    "name": "CDPATH",
+    "description": "A list of directories in which the cd builtin looks for a new directory."
+  },
+  {
+    "name": "fish_term24bit",
+    "description": "If this is set to 1, fish will assume the terminal understands 24-bit RGB color sequences, and won’t translate them to the 256 or 16 color palette. This is often detected automatically."
+  },
+  {
+    "name": "fish_term256",
+    "description": "If this is set to 1, fish will assume the terminal understands 256 colors, and won’t translate matching colors down to the 16 color palette. This is usually autodetected."
+  },
+  {
+    "name": "fish_ambiguous_width",
+    "description": "controls the computed width of ambiguous-width characters. This should be set to 1 if your terminal renders these characters as single-width (typical), or 2 if double-width."
+  },
+  {
+    "name": "fish_emoji_width",
+    "description": "controls whether fish assumes emoji render as 2 cells or 1 cell wide. This is necessary because the correct value changed from 1 to 2 in Unicode 9, and some terminals may not be aware. Set this if you see graphical glitching related to emoji (or other “special” characters). It should usually be auto-detected."
+  },
+  {
+    "name": "fish_autosuggestion_enabled",
+    "description": "controls if Autosuggestions are enabled. Set it to 0 to disable, anything else to enable. By default they are on."
+  },
+  {
+    "name": "fish_handle_reflow",
+    "description": "determines whether fish should try to repaint the commandline when the terminal resizes. In terminals that reflow text this should be disabled. Set it to 1 to enable, anything else to disable."
+  },
+  {
+    "name": "fish_key_bindings",
+    "description": "the name of the function that sets up the keyboard shortcuts for the command-line editor."
+  },
+  {
+    "name": "fish_escape_delay_ms",
+    "description": "sets how long fish waits for another key after seeing an escape, to distinguish pressing the escape key from the start of an escape sequence. The default is 30ms. Increasing it increases the latency but allows pressing escape instead of alt for alt+character bindings. For more information, see the chapter in the bind documentation."
+  },
+  {
+    "name": "fish_sequence_key_delay_ms",
+    "description": "sets how long fish waits for another key after seeing a key that is part of a longer sequence, to disambiguate. For instance if you had bound \\cx\\ce to open an editor, fish would wait for this long in milliseconds to see a ctrl-e after a ctrl-x. If the time elapses, it will handle it as a ctrl-x (by default this would copy the current commandline to the clipboard). See also Key sequences."
+  },
+  {
+    "name": "fish_complete_path",
+    "description": "determines where fish looks for completion. When trying to complete for a command, fish looks for files in the directories in this variable."
+  },
+  {
+    "name": "fish_cursor_selection_mode",
+    "description": "controls whether the selection is inclusive or exclusive of the character under the cursor (see Copy and Paste)."
+  },
+  {
+    "name": "fish_function_path",
+    "description": "determines where fish looks for functions. When fish autoloads a function, it will look for files in these directories."
+  },
+  {
+    "name": "fish_greeting",
+    "description": "the greeting message printed on startup. This is printed by a function of the same name that can be overridden for more complicated changes (see funced)"
+  },
+  {
+    "name": "fish_history",
+    "description": "the current history session name. If set, all subsequent commands within an interactive fish session will be logged to a separate file identified by the value of the variable. If unset, the default session name “fish” is used. If set to an empty string, history is not saved to disk (but is still available within the interactive session)."
+  },
+  {
+    "name": "fish_trace",
+    "description": "if set and not empty, will cause fish to print commands before they execute, similar to set -x in bash. The trace is printed to the path given by the --debug-output option to fish or the FISH_DEBUG_OUTPUT variable. It goes to stderr by default."
+  },
+  {
+    "name": "FISH_DEBUG",
+    "description": "Controls which debug categories fish enables for output, analogous to the --debug option."
+  },
+  {
+    "name": "FISH_DEBUG_OUTPUT",
+    "description": "Specifies a file to direct debug output to."
+  },
+  {
+    "name": "fish_user_paths",
+    "description": "a list of directories that are prepended to PATH. This can be a universal variable."
+  },
+  {
+    "name": "umask",
+    "description": "the current file creation mask. The preferred way to change the umask variable is through the umask function. An attempt to set umask to an invalid value will always fail."
+  },
+  {
+    "name": "BROWSER",
+    "description": "your preferred web browser. If this variable is set, fish will use the specified browser instead of the system default browser to display the fish documentation."
+  }
+]
