+++ src/snippets/statusNumbers.json
@@ -1,0 +1,38 @@
+[
+  {
+    "name": "0",
+    "description": "0 is generally the exit status of commands if they successfully performed the requested operation."
+  },
+  {
+    "name": "1",
+    "description": "1 is generally the exit status of commands if they failed to perform the requested operation."
+  },
+  {
+    "name": "121",
+    "description": "121 is generally the exit status of commands if they were supplied with invalid arguments."
+  },
+  {
+    "name": "123",
+    "description": "123 means that the command was not executed because the command name contained invalid characters."
+  },
+  {
+    "name": "124",
+    "description": "124 means that the command was not executed because none of the wildcards in the command produced any matches."
+  },
+  {
+    "name": "125",
+    "description": "125 means that while an executable with the specified name was located, the operating system could not actually execute the command."
+  },
+  {
+    "name": "126",
+    "description": "126 means that while a file with the specified name was located, it was not executable."
+  },
+  {
+    "name": "127",
+    "description": "127 means that no function, builtin or command with the given name could be located."
+  },
+  {
+    "name": "128",
+    "description": "128 is used when a process exits a signal, plus the number of the signal"
+  }
+]
